requirements.txt
streamlit
plotly
pandas
scikit-learn
numpy
reportlab
openai
kaleido
python-dotenv
prophet